WIP remote build_options to Tools/scripts
